Make chart instances accessible to be waited for via the .ready prop
The new 3.9.0 version of pulumi-kubernetes provides a ".ready" prop on Helm charts. This is meant to be used by resources that are dependent upon the chart. Pulumi will now wait for all resources of the chart to be ready before deploying the dependent resource.

This is helpful for our ingress and aws alb controller deployments. Errors could often occur with the ingress failing to resolve immediately while the controller is still initializing. "skipAwait" was previously used to avoid this issue. Now our ingresses should be able to wait for every resource of the alb chart to be ready before being created.