Set minSize to 2 and remove maxSize
Users should really have 2 instances running so that the 2 tezos nodes can be distributed on them. If a user scales instances to 1 and still keeps 2 tezos nodes, k8s would try and schedule both tezos nodes on the same instance. It would likely result in the volume affinity error. Bec the volume for the 2nd node was created on the second instance which was in a different zone. So this code change makes it more explicit that the minimum number of instances you should have is 2.